requirements/base.txt
psycopg2-binary==2.7.5
python-dateutil==2.7.3
Babel==2.6.0

Django==2.1.1

djangorestframework==3.8.2
django-rest-swagger==2.2.0
django-filter==2.0.0
